% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_advect.R
\name{advectMODPATH}
\alias{advectMODPATH}
\title{Particle Tracking Advection with MODPATH 5}
\usage{
advectMODPATH(mpdt, t1, t2, MFt0, phi_e, disnm, dis, bas, hds, cbb, cbf, newds,
  newcbf, transient, maxnp)
}
\arguments{
\item{mpdt}{data.table; mobile particles}

\item{t1, t2}{numeric [1]; time step start and end}

\item{MFt0}{MODFLOW start time}

\item{phi_e}{porosity}

\item{disnm}{file name for DIS package}

\item{dis}{DIS.MFpackage}

\item{bas}{BAS.MFpackage}

\item{hds}{file name for head save}

\item{cbb}{file name for cell-by-cell budget}

\item{cbf}{CBF file name}

\item{newcbf}{ask MODPATH to write fresh CBF?}

\item{transient}{does the MODFLOW model contain more than one time step?}

\item{maxnp}{number of particles MODPATH should allocate memory for}

\item{newdat}{write fresh DAT?}
}
\value{
data.table with pathline results
}
\description{
Particle Tracking Advection with MODPATH 5
}

