% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_coalesce.R
\name{coalesceDRW}
\alias{coalesceDRW}
\title{Dynamic Coalescing}
\usage{
coalesceDRW(pdt, cd, mm, maxnp, mfdata, wtop, mfts)
}
\arguments{
\item{pdt}{particle data.table (ts, x, y, L, zo, m)}

\item{cd}{numeric [2];
coalescing search radii in the horizontal and vertical directions}

\item{mm}{numeric [1];
minimum mass of particles after coalescing}

\item{maxnp}{integer [1] or \code{Inf};
maximum number of particles after coalescing}

\item{mfdata}{NetCDF object; MODFLOW data set}

\item{wtop}{NetCDF object;
cell by cell top of water (see \code{\link[Rflow]{get.wtop.nc}})}

\item{mfts}{integer [1];
current MODFLOW time step number}
}
\value{
list with:\cr
\code{$state}: particle data.table with column and row references
 (ts, x, y, L, zo, m, C, R)\cr
\code{$loss}: mass lost due to being in inactive cells
}
\description{
Dynamic Coalescing
}

