% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_advect.R
\name{dattxt}
\alias{dattxt}
\title{MODPATH DAT file write}
\usage{
dattxt(por, MXP, dis, bas)
}
\arguments{
\item{por}{numeric; porosity}

\item{MXP}{integer; particles for which memory should be allocated}

\item{dis}{DIS.MFpackage}

\item{bas}{BAS.MFpackage}
}
\value{
character string
}
\description{
MODPATH DAT file write
}

