% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_disperse.R
\name{disperseRW}
\alias{disperseRW}
\title{Random Walk Dispersion}
\usage{
disperseRW(mpdt, D, vdepD, dt, Ndp, sym = TRUE, ThreeDD = FALSE)
}
\arguments{
\item{mpdt}{mobile particles data.table after the advection step
(pno, ts, x, y, L, zo, m, s, traj)}

\item{D}{numeric [2 or 3];
simplified dispersion tensor (L, T, V)}

\item{vdepD}{logical [1];
if \code{TRUE}, \code{D} represents the dispersivity (often denoted
 alpha, units of length), otherwise \code{D} represents the dispersion
 coefficient}

\item{dt}{length of current time step}

\item{Ndp}{integer [1];
number of dispersed particle pairs to spawn from each starting particle}

\item{sym}{logical [1];
if \code{TRUE}, dispersed particle pairs are dispersed symmetrically,
 which ensures that centre of mass is absolutely conserved}

\item{ThreeDD}{logical [1];
not yet coded}
}
\value{
data.table (ts, x, y, L, zo, m)
}
\description{
Random Walk Dispersion
}

