% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_master.R
\name{DRW}
\alias{DRW}
\title{Dynamic Random Walk aqueous contaminant transport solution}
\usage{
DRW(rootname, description, mfdir = ".", mfdata, wtop, dis, bas, wel, hds, cbb,
  cbf, newcbf = TRUE, source.term, STna.rm = FALSE, porosity, start.t,
  end.t, dt, D, vdepD, Rf = 1, lambda = 0, decay.sorbed = FALSE, cd, mm,
  minnp = 100L, maxnp, Ndp = 2L, init = NULL, nc.to.mf = 1L,
  plot.state = TRUE, keep.MF.cellref = TRUE, time.mismatch.tol = 0.001,
  ...)
}
\arguments{
\item{rootname}{character string;
the rootname for the results file (\code{".rds"} is automatically
 affixed)}

\item{description}{character string;
a description of this model run, for future reference}

\item{mfdir}{character string;
path to the directory holding all the MODFLOW input and output files;
 the function navigates to this directory so that extended file paths
 for MODFLOW files needn't be specified}

\item{mfdata}{NetCDF object, list of NetCDF objects or character string[Nmfds];
the MODFLOW data sets in NetCDF form (see \code{\link[Rflow]{GW.nc}}),
 either as NetCDF objects (see \code{\link[RNetCDF]{open.nc}}) or as
 character string file paths; this function may use a series of MODFLOW
 models with identical spatial co-ordinates and grids and with time
 periods that lead on from one another, thus a historic flow model may
 seamlessly be linked to a recent flow model, for example}

\item{wtop}{NetCDF object, list of NetCDF objects or character string [Nmfds];
NetCDF data sets of the cell-by-cell top of water in each cell (the
 lower out of head and cell top); if character string file names are
 used, the files needn't exist beforehand as they can be created using
 \code{\link[Rflow]{get.wtop.nc}}}

\item{dis}{character string [Nmfds];
file paths to DIS package files corresponding to the MODFLOW data sets
 in mfdata (note, this cannot be a list of DIS.MFpackage objects, as the
 file names are needed for MODPATH)}

\item{bas}{BAS.MFpackage object (or list thereof) or character string [Nmfds];
the BAS packages corresponding to mfdata (see
 \code{\link[Rflow]{read.BAS}})}

\item{wel}{Optional: WEL.MFpackage object (or list thereof) or character string
 [Nmfds];
the WEL packages corresponding to mfdata (see
 \code{\link[Rflow]{read.WEL}}); this information is only used for
 plotting, and plotting does not depend on it}

\item{hds, cbb}{character string [Nmfds];
the head save and cell-by-cell budget files corresponding to mfdata;
 although this information is in mfdata, MODPATH needs to know where the
 original HDS files are}

\item{cbf}{character string [Nmfds];
the composite budget file names created by MODPATH 5; if these files
 don't already exist or \code{newcbf = TRUE}, MODPATH 5 will write these
 files afresh}

\item{newcbf}{logical [1];
Whether MODPATH 5 should rewrite the CBF file.  If you have modified the
 MODFLOW model since the last DRW run, then you should set this to
 \code{TRUE} (the default), but this can be time-consuming, so if you
 know it is not needed, set to \code{FALSE}.  The CBF will only ever be
 written for the first time step using a particular MODFLOW data set.}

\item{source.term}{data.table, data.frame, \link[DNAPL]{DNAPLSourceTerm} object, or list of
 any combination of these;
information about the transient point releases in the system; if a data
 table or frame, columns should be: x, y (location), L (MODFLOW layer),
 zo (z-offset within layer) and J (list of functions of one variable,
 time, returning values representing source term flux, in units of mass
 per time, at this point)}

\item{STna.rm}{logical [1];
if \code{TRUE}, any \code{NA} values resulting from the source term
 functions (within the model time frame) are ignored and treated as 0,
 otherwise an error will occur if \code{NA}s are found.}

\item{porosity}{numeric [1], numeric [NLAY] or numeric array [NCOL, NROW, NLAY];
the porosity (fractional, 0 to 1), either as uniform value,
 layer-by-layer values or a 3D array of cell-by-cell values matching
 the MODFLOW grid dimensions}

\item{start.t, end.t}{numeric [1];
start and end times of the model (the \code{\link[td]{td}} function may
 be useful)}

\item{dt}{numeric [1]; time step size}

\item{D}{numeric [2 or 3];
simplified dispersivity/ dispersion coefficient tensor, giving
 longitudinal, transverse and optionally vertical components}

\item{vdepD}{logical [1];
\code{TRUE} if the dispersion coefficient is velocity-dependent, in
 which case \code{D} is taken to represent dispersivity, with units of
 length}

\item{Rf}{numeric [1];
retardation factor (1 - (bulk_density times K_d)/(effective_porosity))}

\item{lambda}{numeric [1]; first-order degradation constant (log(2)/half_life)}

\item{decay.sorbed}{logical [1];
\code{TRUE} if first-order degradation affects sorbed contaminant (not
 yet implemented)}

\item{cd}{numeric [2];
horizontal and vertical coalescing search radii (see
 \code{\link[coalesce]{coalesce}})}

\item{mm}{numeric [1];
minimum mass of particles after coalescing (see
 \code{\link[coalesce]{coalesce}})}

\item{minnp}{integer [1];
minimum number of particles on which to perform coalescing}

\item{maxnp}{integer [1];
maximum number of particles after coalescing  (see
 \code{\link[coalesce]{coalesce}})}

\item{Ndp}{integer [1];
number of pairs of particles to spawn to model the dispersion of
 contaminant mass}

\item{init}{\code{NULL}, a \link{DRWmodel} object or a character string giving the
 RDS file name in which a DRWmodel object is stored;
another DRW model run to use as the initial state for the current model;
 the time step just before \code{start.t} is used as the initial state,
 modifying the time step times as necessary}

\item{nc.to.mf}{integer [\code{length(mfdata)}];
the MODFLOW model that is related to by each NetCDF dataset in
 \code{mfdata}; normally this will be 1, 2, 3 ..., but sometimes a
 large MODFLOW model may be split into multiple NetCDF datasets (see
 \code{\link[Rflow]{GW.nc}}), in which case it will be something like
 1, 1, 1 ...; for the most common case in which there is only one
 MODFLOW model used, the default value of 1 is appropriate}

\item{plot.state}{logical [1];
if \code{TRUE}, a summary plot of the model state will be shown after
 each time step (generally, not costly in terms of run time)}

\item{keep.MF.cellref}{logical [2];
if \code{TRUE}, the final results for mobile and immobile particles will
 retain MODFLOW column (C), row (R), dataset (mfds) and time step (mfts)
 references}

\item{time.mismatch.tol}{numeric [1];
various time comparisons are made during the model in order to check
 consistency, such as checking that the end time of one MODFLOW dataset
 matches the start time of the next; time.mismatch.tol gives the
 tolerance used in these comparisons in calls to
 \code{\link[base]{all.equal}}}

\item{...}{graphical parameters such as \code{xlim} (not axis labels, titles,
\code{asp} or \code{zlim})}
}
\value{
A \link{DRWmodel} object, invisibly.  The result is also saved to file:
 \code{paste0(rootname, ".rds")}.
}
\description{
Dynamic Random Walk aqueous contaminant transport solution
}
\examples{
library(data.table)

# single point source which ceases after t = 10000
mfdir <- system.file(package = "DRW")
demoDRW <- DRW("DRW_EXAMPLE", "demo", mfdir,
               "drw_mf_demo.nc", "drw_mf_demo_wtop.nc",
               "drw_mf_demo.dis", "drw_mf_demo.bas", "drw_mf_demo.wel",
               "drw_mf_demo.hds", "drw_mf_demo.cbb", "DRWtest.cbf",
               newcbf = TRUE,
               source.term = data.table(x = 625, y = 825,
                                        L = 1L, zo = .5,
                                        J = function(t){
                                          if(t < 1e4) 1 else 0
                                        }),
               porosity = .1,
               start.t = 9000, end.t = 11000, dt = 200,
               D = c(10, 1), vdepD = TRUE,
               cd = c(20, 10), mm = 1e-7, minnp = 100L, maxnp = 2e4L,
               Ndp = 4L)

}

