% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_timestep.R
\name{DRW_timesteps}
\alias{DRW_timesteps}
\title{Set up time steps for DRW model}
\usage{
DRW_timesteps(start.t, end.t, dt, dsett, time.mismatch.tol)
}
\arguments{
\item{start.t}{numeric [1];
start and end times of the model (the \code{\link[td]{td}} function may
 be useful)}

\item{end.t}{numeric [1];
start and end times of the model (the \code{\link[td]{td}} function may
 be useful)}

\item{dt}{numeric [1]; time step size}

\item{dsett}{numeric [];
MODFLOW dataset time divides, including start and end}

\item{time.mismatch.tol}{numeric [1];
various time comparisons are made during the model in order to check
 consistency, such as checking that the end time of one MODFLOW dataset
 matches the start time of the next; time.mismatch.tol gives the
 tolerance used in these comparisons in calls to
 \code{\link[base]{all.equal}}}
}
\value{

}
\description{
Set up time steps for DRW model
}

