% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_master.R
\docType{class}
\name{DRWmodel-class}
\alias{DRWmodel}
\alias{DRWmodel-class}
\title{DRW model results formal class}
\value{
DRWmodel S4 object
}
\description{
DRW model results formal class
}
\section{Slots}{

\describe{
\item{\code{time}}{numeric;
time values of the time step divides, including start and end of model}

\item{\code{plume}}{data.table with columns:\cr
\code{$ts} (int, key): DRW time step\cr
\code{$x, $y, $z} (num): 3D position\cr
\code{$L} (int): MODFLOW layer\cr
\code{$zo} (num): z-offset within layer (0: cell base, to 1: top of
 groundwater in cell (see \code{\link[Rflow]{get.wtop.nc}}))\cr
\code{$m} (num): particle mass;
mobile particle swarm}

\item{\code{sorbed}}{data.table, (columns as for \code{plume});
immobile particles}

\item{\code{release}}{data.table with columns:\cr
\code{$x, $y, $L, $zo}: as with \code{plume}\cr
\code{$J} (list): list of functions giving flux in terms of time\cr
the source term release at each source location}

\item{\code{fluxout}}{data.table with columns:\cr
\code{$ts} (int, key): DRW time step\cr
\code{$C, $R, $L} (int): MODFLOW cell reference\cr
\code{$J_out} (num): rate of mass loss in cell during this time step\cr
outflux of mass from particles by MODFLOW cell reference and DRW time
 step}

\item{\code{lost}}{matrix [ndrts, 7];
mass lost from model other than in sinks; each named column refers to
 mass lost from each edge of the MODFLOW model (top, left, right, bottom),
 failed source release due to release into an inactive or dry MODFLOW
 cell, degraded in first-order reactions, or other}

\item{\code{dispersion}}{list with elements:\cr
\code{$D} named numeric [2 or 3]: longitudinal, transverse (and vertical)
 dispersivities (units of length) if \code{vdepD} or dispersion
 coefficients (units of length^2/time) otherwise\cr
\code{$vdepD} logical [1]: see above\cr
\code{$`3D`} logical [1]: \code{TRUE} if a 3D dispersion tensor is used}

\item{\code{reactions}}{list with elements:\cr
\code{$Rf} numeric [1]: retardation factor\cr
\code{$lambda} numeric [1]: first-order reactive decay constant
 (equivalent to log(2)/half-life)\cr
\code{$decaysorbed} logical [1]: whether first-order reactive decay
 applies to immobile solute mass}

\item{\code{porosity}}{numeric array [1], [NLAY] or [NCOL, NROW, NLAY];
effective porosity of each MODFLOW cell, either uniform, by layer or by
cell}

\item{\code{coalescing}}{list with elements:\cr
\code{$cd} named numeric [2]: coalescing search radii horizontally and
 vertically\cr
\code{$mm} numeric [1]: minimum mass of particles after coalescing\cr
\code{$maxnp} integer [1]: maximumn number of particles after
 coalescing\cr
see (\code{\link[coalesce]{coalesce}}) for more details}

\item{\code{description}}{character string;
description of the model run for future reference}

\item{\code{MFinfo}}{character string matrix[number of MODFLOW models, 3];
descriptive attributes, including date, of the MODFLOW NetCDFs, for
 future reference, so that it is clear what flow field (and what version)
 was used in the model}

\item{\code{run.timings}}{POSIXct [];
summary of simulation timings, to give date of model run and for
 performance analysis}
}}

