% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksConc.R
\name{ksConc}
\alias{ksConc}
\title{Weighted Kernel Density Estimate of Concentration}
\usage{
ksConc(DRWmodel, mfdata, wtop, dkcell, ts = NULL, L = NULL, smd, H,
  binned = NA, Kxlim = c("auto", "mf")[1L], Kylim = NULL, ...,
  ptype = "plume")
}
\arguments{
\item{DRWmodel}{\code{DRWmodel} S4 object or a character string giving the file path to
one}

\item{mfdata, wtop}{open NetCDF, character string file path(s) or list of open NetCDFs;
MODFLOW data set(s) and saruated water tops}

\item{dkcell}{numeric [1] or [2];
cell spacing for the kernel smooth in x and y directions}

\item{ts, L}{\code{NULL} or integer[];
time steps/ layers to process (\code{NULL} implies all)}

\item{smd}{numeric [1];
smoothing distance for kernel smooth: too small and the result will be
 patchy, too large and the result will be too spread}

\item{H}{matrix [2, 2];
alternative, lower level way of specifying smoothing (see
 \code{\link[ks]{kde}})}

\item{binned}{logical [1];
see \code{\link[ks]{kde}}; \code{NA} uses binning for particle sets
 numbering more than 500 and not otherwise}

\item{Kxlim}{\code{"auto"}, \code{"mf"}, numeric [2] or something that, with
 \code{Kylim}, can be read by \code{\link[grDevices]{xy.coords}};
\code{Kxlim} and \code{Kylim} give the x and y limits for the region in
 which to determine the concentration; "auto" uses the particle
 distribution and "mf" uses the MODFLOW model edges}

\item{Kylim}{\code{NULL} or numeric [2]}

\item{...}{additional arguments for \code{\link[ks]{kde}}}

\item{ptype}{\code{"plume"} (the default) or \code{"sorbed"}}
}
\value{
List object of class kDRW with elements:\cr
\code{$conc} numeric [x, y, L, ts]; concentration estimate\cr
\code{$coords} list:\cr
\code{..$x,y} numeric []; x and y grid divide co-ordinates (note not the
 cell centres)\cr
\code{..$L,ts} integer []; layers and time steps that are present\cr
\code{$H} numeric [2, 2]; the H matrix used for smoothing (see
 \code{\link[ks]{kde}})
}
\description{
A post-processing tool for \code{\link{DRW}}.  This function gives a
 distributed concentration estimate for a DRW result, using a weighted
 kernel density estimate from the particles, using
 \code{\link[ks]{kde}}.
}
\examples{
setwd(mfdir <- system.file(package = "DRW"))
drmod <- readRDS("DRW_EXAMPLE.rds")

kdr <- ksConc(drmod, "drw_mf_demo.nc", "drw_mf_demo_wtop.nc", 10,
              smd = 20, Kxlim = "mf")

# Note here that the result is indexed using strings for the 3rd and 4th
#  dimensions.  This is much less likely to cause confusion because the
#  layers and time steps included are likely not to be a simple sequence
#  starting from 1.
with(kdr, image(z = conc[,, "L1", "ts10"], coords$x, coords$y,
                col = grDevices::rainbow(51L)))

}

