% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_advect.R
\name{namtxt}
\alias{namtxt}
\title{MODPATH name file text}
\usage{
namtxt(dis, hds, cbb, dat)
}
\arguments{
\item{dis, hds, cbb, dat}{file names, not too long (use local names)}
}
\value{
character string
}
\description{
MODPATH name file text
}

