% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nc.imtx}
\alias{nc.imtx}
\title{Index matrix for NetCDF}
\usage{
nc.imtx(ncfile, variable, imtx)
}
\arguments{
\item{ncfile}{NetCDF object (see \code{\link[RNetCDF]{open.nc}})}

\item{variable}{character string; the variable name in the NetCDF file}

\item{imtx}{integer matrix;
the indices to read; must have the same number of columns as variable
 has dimensions; must contain only positive integers and no NAs; ideally
 the range of values in some or most columns would not be very large to
 avoid reading a large array into working memory}
}
\value{
a vector whose type corresponds to that of \code{variable}
}
\description{
Emulate the \code{A[B]} matrix indexing behaviour in R for NetCDF
 connections whilst avoiding reading large arrays into memory.
}

