% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_plot.R
\name{plotDRWstate}
\alias{plotDRWstate}
\title{Plot DRW state on go}
\usage{
plotDRWstate(mpdt, rpdt, drts, mfsp, bas, wel, gccs, grcs, ...)
}
\arguments{
\item{mpdt}{mobile particle data table (x, y, L)}

\item{rpdt}{source release data table (x, y, L)}

\item{drts}{integer [1]; DRW time step}

\item{mfsp}{integer [1]; MODFLOW stress period}

\item{bas}{BAS.MFpackage}

\item{wel}{WEL.MFpackage or NULL}

\item{gccs, grcs}{numeric []; column and row divider co-ordinates}

\item{...}{graphical parameters such as \code{xlim} (not axis labels, titles,
\code{asp} or \code{zlim})}
}
\value{
\code{NULL}
}
\description{
Plot DRW state on go
}

