% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_advect.R
\name{ptrtxt}
\alias{ptrtxt}
\title{MODPATH particle starting locations}
\usage{
ptrtxt(ptr.dat, rt)
}
\arguments{
\item{ptr.dat}{data.table (x, y, L, zo)}

\item{rt}{numeroc [1]; release time}
}
\value{
character string
}
\description{
MODPATH particle starting locations
}

