% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_disperse.R
\name{Rot}
\alias{Rot}
\title{2D rotation matrix}
\usage{
Rot(phi, scale = 1)
}
\arguments{
\item{phi}{numeric [1]; angle in radians}

\item{scale}{numeric [1]; expansion factor}
}
\value{
2 by 2 matrix
}
\description{
2D rotation matrix
}

