% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_sorb.R
\name{sorb.desorb}
\alias{sorb.desorb}
\title{Sorb and Desorb}
\usage{
sorb.desorb(mpdt, ipdt, Rf)
}
\arguments{
\item{mpdt, ipdt}{data.table (ts, x, y, L, zo, m);
for the mobile and immobile states respectively}

\item{Rf}{numeric;
retardation factor, conceptually intended to be >= 1}
}
\value{
named list of data.tables for the new mobile and immobile states
}
\description{
Sorb and Desorb
}

