% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_source.R
\name{ST.DNAPL}
\alias{ST.DNAPL}
\title{DRW source term from a DNAPL source term}
\usage{
ST.DNAPL(dnst, mfdata, gccs, grcs)
}
\arguments{
\item{dnst}{\link[DNAPL]{DNAPLSourceTerm} object}

\item{mfdata}{list of NetCDFs of MODFLOW output}

\item{gccs, grcs}{numeric[];
column and row divider co-ordinates}
}
\value{
data.table, source term for \code{\link{DRW}}
}
\description{
see \code{\link[DNAPL]{DNAPLSourceTerm}}
}

