% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{DRWwell-analysis}
\alias{DRWwell-analysis}
\alias{wellC}
\alias{wellJ}
\title{Analyse receptors in a DRW model output}
\usage{
wellJ(dr, welref)

wellC(dr, WEL, DIS, MFt0, welref = WEL)
}
\arguments{
\item{dr}{a \link{DRWmodel} S4 object,}

\item{welref}{data.table (or object coercible to one) with named columns C, R and L
 giving the CRL references of receptors to be analysed, integer [3],
 \code{NULL}, WEL.MFpackage (see \code{\link[Rflow]{read.WEL}}), or (for
 \code{wellJ} only) various other MFpackage objects;
\code{NULL} implies all cell references to which a flux has been recorded
 by \code{\link{DRW}}}

\item{WEL}{WEL.MFpackage or NetCDF object of MODFLOW outputs;
if a NetCDF object, there must be a "Wells" array in the data set}

\item{DIS}{DIS.MFpackage (see \code{\link[Rflow]{read.DIS}});
only required if \code{WEL} is a WEL.MFpackage object}

\item{MFt0}{numeric [1];
only required if \code{WEL} is a WEL.MFpackage object
start time of the MODFLOW model, which should be the same as for the
 NetCDF MODFLOW data set used for the \code{\link{DRW}} model}
}
\value{
\code{wellJ}: list of fluxes to each receptor specified in welref, with
 a value for each time step in \code{dr} (\code{dr@time})\cr
\code{wellC}: as for \code{wellJ} but concentrations, determined by
 diluting the flux with the transient abstraction rate

In each case, the elements of the list are named giving the cell
 reference as a character string (e.g. \code{"C30R20L1"})
}
\description{
Extract the flux to individual receptors and calculate the abstracted
 concentration to receptors in the output of a \code{\link{DRW}} model.
}

